added cmake files for easy inclusion for cmake projects
