+++ avtk/utils.cxx
@@ -35,41 +35,9 @@
 #include "theme.hxx"
 #include <algorithm>
 #include <cstring>
-#include <sndfile.h>
 
 namespace Avtk
 {
-int loadSample( std::string path, std::vector< float >& sample, bool printErrors )
-{
-#ifdef AVTK_SNDFILE
-	SF_INFO info;
-	memset( &info, 0, sizeof( SF_INFO ) );
-	SNDFILE* const sndfile = sf_open( path.c_str(), SFM_READ, &info);
-	if ( !sndfile ) {
-		printf("Failed to open sample '%s'\n", path.c_str() );
-		return -1;
-	}
-
-	if( !(info.channels == 1 || info.channels == 2) ) {
-		int chnls = info.channels;
-		printf("Loading sample %s, channels = %i\n", path.c_str(), chnls );
-		return -1;
-	}
-
-	sample.resize( info.frames * info.channels );
-
-	sf_seek(sndfile, 0ul, SEEK_SET);
-	sf_read_float( sndfile, &sample[0], info.frames * info.channels );
-	sf_close(sndfile);
-
-	return OPENAV_OK;
-#else
-	if( printErrors ) {
-		printf("AVTK compiled without SNDFILE support: cannot load audio sample.\n");
-	}
-	return OPENAV_ERROR;
-#endif
-}
 
 int fileUpLevel( std::string path, std::string& newPath )
 {
